+++ src/main/kotlin/io/github/t45k/slidekt/api/AspectRatio.kt
@@ -1,4 +1,4 @@
-package io.github.t45k.kmp_trial.api
+package io.github.t45k.slidekt.api
 
 enum class AspectRatio(val ratio: Float) {
     RATIO_16_9(16f / 9f), RATIO_4_3(4f / 3f);
